<refactor>[kvmagent]: move embedded services to plugins/services/
Relocate 4 embedded service files from plugins/ to plugins/services/:
prometheus.py, ft_vm_fencer.py, gpu_service_manager.py, qga_zwatch.py

PluginRegistry uses os.walk() so runtime behavior is unchanged.
mini_fencer.py stays in plugins/ as it is a library, not a service.
Update 2 test file imports to kvmagent.plugins.services.

Resolves: ZCF-0
